% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abind_2.R
\name{abind.2}
\alias{abind.2}
\title{Combine multi-dimensional arrays}
\usage{
abind.2(..., along = N, rev.along = NULL, new.names = NULL,
  force.array = TRUE, make.names = use.anon.names, use.anon.names = FALSE,
  use.first.dimnames = FALSE, hier.names = FALSE, use.dnns = FALSE)
}
\arguments{
\item{...}{a}

\item{along}{a}

\item{rev.along}{a}

\item{new.names}{a}

\item{force.array}{a}

\item{make.names}{a}

\item{use.anon.names}{a}

\item{use.first.dimnames}{a}

\item{hier.names}{a}

\item{use.dnns}{a}
}
\description{
Combine multi-dimensional arrays. This is a generalization of cbind and rbind. Takes a sequence of vectors, matrices,
}
\author{
Tony Plate <tplate@acm.org> and Richard Heiberger
}

