% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_to_data_frame.R
\name{ar.to.df}
\alias{ar.to.df}
\title{Convert landamrk array into data frame}
\usage{
ar.to.df(data, specimens.names = FALSE)
}
\arguments{
\item{data}{landmark array}

\item{specimens.names}{optional vector containing specimens names}
}
\description{
Converts landmark array into data frame with additional column with specimens name. Can be used with lattice xyplot function to plot all specimens.
}

