% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_tps_with_variables.R
\name{bind.tps.with.variables}
\alias{bind.tps.with.variables}
\title{Bind landmark arrays}
\usage{
bind.tps.with.variables(..., list.of.variables, as.factor = TRUE)
}
\arguments{
\item{...}{arrays to bind}

\item{list.of.variables}{list of variables, each with levels corresponding to number of binded arrays}

\item{as.factor}{logical indicating whether variables should be returned as factor}
}
\description{
Returns list with, as first element, binded landmark arrays readed from the tps files and, as next elements, 
variables (taxon names, localisations, etc.).
}
