% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_curve.R
\name{change.curves}
\alias{change.curves}
\title{Remove landmarks from semilandmark matrix}
\usage{
change.curves(curves, points.to.remove)
}
\arguments{
\item{curves}{Semilandmark matrix}

\item{points.to.remove}{Vector of landamrk numbers to be removed}
}
\description{
Function removes landmarks from semilandmark matrix and changes numbers of remining ones.
}
\examples{
change.curves()
}

