% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_id.R
\name{clear.id}
\alias{clear.id}
\title{Clear specimen ID}
\usage{
clear.id(string.to.clean, split = "\\\\")
}
\arguments{
\item{string.to.clean}{specimens ID}

\item{split}{split pattern (directory mark); default = '\\'}
}
\description{
Clears specimens ID from directory and file extension.
}

