% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_specimens_names.R
\name{clear.specimens.names}
\alias{clear.specimens.names}
\title{Clear specimens ID's in landmark array}
\usage{
clear.specimens.names(data, split = "\\\\")
}
\arguments{
\item{data}{landmark array}

\item{split}{split pattern (directory mark); default = '\\'}
}
\description{
Clears specimens ID's in landmark array from directory and file extension.
}
