% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_nr.R
\name{convert.nr}
\alias{convert.nr}
\title{Convert landmark number}
\usage{
convert.nr(landmark, mid, left, right, asym.to.sym = TRUE)
}
\arguments{
\item{landmark}{Landmark number}

\item{mid}{Midline landmarks}

\item{left}{Leftside landmark}

\item{right}{Righside landmarks}

\item{asym.to.sym}{Direction of conversion. Default from asymetric to symmetrized}
}
\description{
A function to convert landmark number between asymmetric and symmetrized version.
}
\examples{
convert.nr()
}
\keyword{conversion}
\keyword{landmarks,}

