% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_vector.R
\name{convert.nr.vector}
\alias{convert.nr.vector}
\title{Convert multiple landmark numbers}
\usage{
convert.nr.vector(landmark.vector, mid, left, right, asym.to.sym = TRUE)
}
\arguments{
\item{landmark.vector}{Vector of landmark numbers}

\item{mid}{Midline landmarks}

\item{left}{Leftside landmark}

\item{right}{Righside landmarks}

\item{asym.to.sym}{Direction of conversion. Default from asymetric to symmetrized}
}
\description{
Function to convert a set of landmark numbers between asymmetric and symmetrized version.
}
\keyword{conversion}
\keyword{landmarks,}

