% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_parameters.R
\name{count.parameters}
\alias{count.parameters}
\title{Count linear parameters}
\usage{
count.parameters(data, parameters.list, transverse = NULL, saggital = NULL)
}
\arguments{
\item{data}{landmark matrix}

\item{parameters.list}{list of parameters - name of element should be the name of parameter and it has to contain 
landmark numbers (2 or more)}

\item{transverse}{names of transverse parameters (if they are counted from groups of landmarks)}

\item{saggital}{names of saggital parameters (if they are counted from groups of landmarks)}
}
\description{
Counts linear parameters from landmark data. If elemens are specified as group of landmarks, object should be rotated
 - saggital axis should be parallel to y axis.
}

