% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deformacja.R
\name{deformacja}
\alias{deformacja}
\title{Deform landmark data}
\usage{
deformacja(data, strain.matrix)
}
\arguments{
\item{data}{Landmark data (single specimens matrix)}

\item{strain.matrix}{Strain matrix}
}
\description{
Deforms landmark data using strain matrix.
}

