% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_curves.R
\name{find.curves}
\alias{find.curves}
\title{Find curves}
\usage{
find.curves(curves)
}
\arguments{
\item{curves}{Matrix of curves}
}
\description{
Finds beginings of curves in semilandmark curves matrix used in geomorph::gpagen
}
