% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_landmarks.R
\name{find.missing.landmarks}
\alias{find.missing.landmarks}
\title{Replace missing landmarks in array}
\usage{
find.missing.landmarks(data, clear.names = FALSE, split = "\\\\")
}
\arguments{
\item{data}{landmark array}

\item{clear.names}{optional bolean for clearing specimens IDs; default = FALSE}

\item{split}{optional split pattern for clearing IDs (directory mark); default = '\\'}
}
\description{
Replaces missing landmarks in array with NA. Based on k-means clustering, seeks for separated group of landmarks (for example you can put all missing landmarks in the corner of the picture). Also handles with tpsdig way of writing missing andmarks (-1, -1). Also can clear specimens IDs.
}

