% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_pairs.R
\name{find.missing.pairs}
\alias{find.missing.pairs}
\title{Find missing pairs in array}
\usage{
find.missing.pairs(data, right, left)
}
\arguments{
\item{data}{landmark array}

\item{right}{right side landmarks}

\item{left}{left side landmarks}
}
\description{
Find missing pairs of landmarks in array.
}

