% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullprocdist.R
\name{fpdist}
\alias{fpdist}
\title{Procrustes distance}
\usage{
fpdist(x, y, reflect = FALSE)
}
\arguments{
\item{x}{Object 1}

\item{y}{Object 2}

\item{reflect}{reflect}
}
\description{
Finds Procrustes distance between two objects
}

