% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_shapes.R
\name{get.mean.shapes}
\alias{get.mean.shapes}
\title{Calculate mean shape for size classes}
\usage{
get.mean.shapes(data, Csize, breaks, simplify = TRUE)
}
\arguments{
\item{data}{landmark array}

\item{Csize}{size vector}

\item{breaks}{number of intervals or specified cutpoints}

\item{simplify}{does output should be a data frame? deafult = TRUE}
}
\description{
Calculates ean shapes for size classes.
}
