% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_matrix.R
\name{list.to.matrix}
\alias{list.to.matrix}
\title{Converts a list of semilandmarks into matrix}
\usage{
list.to.matrix(curves.list)
}
\arguments{
\item{curves.list}{List of semilandmark vectors}
}
\description{
Function converts a list of semilandmarks into a matrix that can be used in geomorph::gpagen.
}
