% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_list.R
\name{matrix.to.list}
\alias{matrix.to.list}
\title{Convert semilandmark matrix into list}
\usage{
matrix.to.list(curves, break.points)
}
\arguments{
\item{curves}{Semilandmark matrix}

\item{break.points}{First rows of the curves}
}
\description{
Function converts semilandmark matrix into list form, each curve in the separate element of the list.
}
