% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_landmarks.R
\name{missing.landmarks}
\alias{missing.landmarks}
\title{Replace missing landmarks}
\usage{
missing.landmarks(data)
}
\arguments{
\item{data}{landmark matrix}
}
\description{
Replaces missing landmarks with NA. Based on k-means clustering, seeks for separated group of landmarks (for example you can put all missing landmarks in the corner of the picture). Also handles with tpsdig way of writing missing andmarks (-1, -1).
}

