% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pairs.R
\name{missing.pairs}
\alias{missing.pairs}
\title{Find missing pairs in matrix}
\usage{
missing.pairs(data, right, left)
}
\arguments{
\item{data}{landmark matrix}

\item{right}{right side landmarks}

\item{left}{left side landmarks}
}
\description{
Find missing pairs of landmarks in matrix.
}

