% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSymm.R
\name{OSymm}
\alias{OSymm}
\title{Symmetrize landmarks}
\usage{
OSymm(X, midline, right, left)
}
\arguments{
\item{X}{Landmark set}

\item{right}{Righside landmarks}

\item{left}{Leftside landmark}

\item{mid}{Midline landmarks}
}
\description{
Symmetrize and estimate missing landmarks in bilateral set.
}
\examples{
OSymm()
}
\author{
Annat Haber
}

