% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters_data.R
\name{parameters.data}
\alias{parameters.data}
\title{Create parameters data frame}
\usage{
parameters.data(data, parameters.list, transverse = NULL, saggital = NULL,
  taxon = NULL)
}
\arguments{
\item{data}{landmark array}

\item{parameters.list}{list of parameters - name of element should be the name of parameter and it has to contain 
landmark numbers (2 or more)}

\item{transverse}{names of transverse parameters (if they are counted from groups of landmarks)}

\item{saggital}{names of saggital parameters (if they are counted from groups of landmarks)}

\item{taxon}{optional taxon name}
}
\description{
Creates data frame containing parameters from landmark array.
}

