% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_landmarks_from_vector.R
\name{remove.landmarks.from.vector}
\alias{remove.landmarks.from.vector}
\title{Remove landmarks from vector}
\usage{
remove.landmarks.from.vector(vector, landmarks)
}
\arguments{
\item{vector}{vector}

\item{landmarks}{landmarks to remove}
}
\description{
Removes landmarks from vector.
}

