% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_landmarks.R
\name{remove.landmarks}
\alias{remove.landmarks}
\title{Remove selected landmarks}
\usage{
remove.landmarks(data, landmarks.to.remove, curves = NULL)
}
\arguments{
\item{data}{single array or a list of arrays with landmark data}

\item{landmarks.to.remove}{vector of landmark numbers to remove}

\item{curves}{optional curves matrix}
}
\description{
Removes selected landmarks from a list of arrays and from curves matrix.
}

