% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_landmark_from_curve.R
\name{rm.landmark.from.curve}
\alias{rm.landmark.from.curve}
\title{Remove selected landmark}
\usage{
rm.landmark.from.curve(curve, landmark.nr)
}
\arguments{
\item{curve}{Curve to change}

\item{landmark.nr}{Number of landmark to remove}
}
\description{
Function removes selected landmark from a curve and changes numbers of remaining ones
}
\examples{
rm.landmark.from.curve 
}

