% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proporcje_kranidium.R
\name{shape.proportions}
\alias{shape.proportions}
\title{Shape proportions}
\usage{
shape.proportions(coords, sag.tr = TRUE)
}
\arguments{
\item{coords}{landmark array}

\item{sag.tr}{proportions direction (default saggital/transversal)}
}
\description{
Calculate shape proportions. Shape axes shoud be parallel to x-y coordinate axes.
}

