% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symetria_2.R
\name{symetria.2}
\alias{symetria.2}
\title{Symmetrization of bilateral landmarks dataset}
\usage{
symetria.2(dane, mid, left, right)
}
\arguments{
\item{dane}{A set of specimens}

\item{mid}{Midline landmarks}

\item{left}{Leftside landmark}

\item{right}{Righside landmarks}
}
\description{
Symmetrize and estimate missing landmarks in a set of specimens.
}
\examples{
symetria.2()
}

