% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector_to_matrix.R
\name{vector.to.matrix}
\alias{vector.to.matrix}
\title{Convert vector of semilandmarks into matrix}
\usage{
vector.to.matrix(single.curve)
}
\arguments{
\item{single.curve}{Vector to be converted}
}
\description{
Function converts vector of semilandmarks into matrix version
}

